refactor(core): split bind contracts, scene assembly, and layout helpers
Extract geom/stat type contracts, required channels, and color-on-fill
warnings from bind-layer-geom-contracts; geometry batches and scene
build from assemble-scene; layout constants/formatters/dedupe from
layout-helpers; and backends/fields from layer-contracts. Facades keep
import paths stable. Add layout-helpers characterization tests.
