+++ packages/core/src/pipeline/layout-constants.ts
@@ -1,0 +1,9 @@
+/**
+ * Shared layout band and legend spacing constants for pipeline chrome.
+ */
+export const TITLE_BAND = 22;
+export const SUBTITLE_BAND = 16;
+export const CAPTION_BAND = 14;
+export const AXIS_TITLE_BAND = 18;
+export const LEGEND_GAP = 12;
+export const LEGEND_EDGE_PAD = 2;
